api.bookfarm/requirements.txt
asgiref==3.2.3
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
dj-database-url==0.5.0
Django==3.0.3
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
lxml==4.5.0
mysqlclient==1.4.6
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.23.0
service-identity==18.1.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
txaio==20.4.1
urllib3==1.25.8
whitenoise==5.0.1
wincertstore==0.2
zope.interface==5.0.2